+++ tools/opencv_capture_benchmark.py
@@ -1,6 +1,6 @@
-device = 2 # This determines the video capture device that will be used for the benchmark.
+device = 0 # This determines the video capture device that will be used for the benchmark.
 resolution = [1280, 720] # This determines the resolution that will be used for the video capture device.
-frames = 240 # This determines how many frames will be captured for the benchmark.
+frames = 400 # This determines how many frames will be captured for the benchmark.
 
 
 import cv2